LayoutTests/platform/gtk/accessibility/press-targets-center-point-expected.txt

target
Actual: (89, 17)
Expected: (89, 18)

translate transform target
Actual: (289, 264)
Expected: (289, 265)

vertical left-to-right writing mode target
Actual: (17, 307)
Expected: (18, 307)

right-to-left direction target
Actual: (129, 509)
Expected: (129, 510)

This tests that press targets the receiving elements center point.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS successfullyParsed is true

TEST COMPLETE
